Re-organises into a more transparent, factored file structure.
