+++ easylist.vcxproj
@@ -145,6 +145,8 @@
   <ItemGroup>
     <ClInclude Include="convstr.h" />
     <ClInclude Include="easylist.h" />
+    <ClInclude Include="easylist_util.h" />
+    <ClInclude Include="object_list.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
