+++ easylist.vcxproj.filters
@@ -26,5 +26,11 @@
     <ClInclude Include="convstr.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="object_list.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="easylist_util.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
