+++ easylist_util.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#define EASYLIST_TEMPLATE_OBJ_LIST_DEFN template<class _Type, class _Alloc = std::allocator<_Type>>
+#define EASYLIST_TEMPLATE_OBJ_LIST_DECL template<class _Type, class _Alloc>
